% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchRequests.R
\name{batch.admin.directory.v1}
\alias{batch.admin.directory.v1}
\title{Makes a Batch API Request to the Admin SDK API}
\usage{
batch.admin.directory.v1(
  requests,
  content_ids,
  gargle_token = gargle::token_fetch(),
  base_url = "https://admin.googleapis.com",
  batch_path = "batch/admin/directory/v1",
  batch_delim = "gpm_batch_delimiter",
  return_response = F
)
}
\arguments{
\item{requests}{A list of requests to the same endpoint based on generated functions from this package. Set return_request = TRUE for the function calls.}

\item{content_ids}{A list of the same length as requests with identifiers for the data returned. These will be the names of the list of results when setting return_response = FALSE.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{base_url}{The base URL for the API. Defaults to 'https://admin.googleapis.com'.}

\item{batch_path}{Defaults to 'batch/admin/directory/v1'. If you get a 404 error, you will need to consult the \href{https://developers.google.com/admin-sdk/}{API documentation} to find the correct batch path.}

\item{batch_delim}{A string used to separate requests in the batch request body. Defaults to 'gpm_batch_delimiter'.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{package_make}}.
}
\details{
The server's response is a single standard HTTP response with a multipart/mixed content type; each part is the response to one of the requests in the batched request, in the same order as the requests.
}
\seealso{
\href{https://developers.google.com/gmail/api/guides/batch#details}{Sample Gmail Batch Request Info}
}
