% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromeosdevices.R
\name{chromeosdevices.update}
\alias{chromeosdevices.update}
\title{Updates a device's updatable properties, such as \code{annotatedUser}, \code{annotatedLocation}, \code{notes}, \code{orgUnitPath}, or \code{annotatedAssetId}.}
\usage{
chromeosdevices.update(
  deviceId,
  customerId,
  projection = NULL,
  screenshotFiles = NULL,
  firmwareVersion = NULL,
  systemRamFreeReports = NULL,
  orderNumber = NULL,
  ethernetMacAddress = NULL,
  platformVersion = NULL,
  deviceFiles = NULL,
  macAddress = NULL,
  willAutoRenew = NULL,
  recentUsers = NULL,
  annotatedLocation = NULL,
  systemRamTotal = NULL,
  status = NULL,
  annotatedUser = NULL,
  model = NULL,
  supportEndDate = NULL,
  ethernetMacAddress0 = NULL,
  notes = NULL,
  serialNumber = NULL,
  diskVolumeReports = NULL,
  cpuStatusReports = NULL,
  bootMode = NULL,
  orgUnitPath = NULL,
  lastSync = NULL,
  annotatedAssetId = NULL,
  osVersion = NULL,
  lastEnrollmentTime = NULL,
  kind = NULL,
  lastKnownNetwork = NULL,
  activeTimeRanges = NULL,
  manufactureDate = NULL,
  tpmVersionInfo = NULL,
  meid = NULL,
  dockMacAddress = NULL,
  etag = NULL,
  autoUpdateExpiration = NULL,
  callback = NULL,
  prettyPrint = NULL,
  alt = NULL,
  quotaUser = NULL,
  fields = NULL,
  uploadType = NULL,
  upload_protocol = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.chromeosdevices.update"]]$scopes[[1]]),
  return_response = F
)
}
\arguments{
\item{deviceId}{The unique ID of the device. The \code{deviceId}s are returned in the response from the \href{/admin-sdk/v1/reference/chromeosdevices/list}{chromeosdevices.list} method.}

\item{customerId}{The unique ID for the customer's Google Workspace account. As an account administrator, you can also use the \code{my_customer} alias to represent your account's \code{customerId}. The \code{customerId} is also returned as part of the \href{/admin-sdk/directory/v1/reference/users}{Users resource}.}

\item{projection}{Restrict information returned to a set of selected fields.}

\item{screenshotFiles}{List of screenshot files to download. Type is always "SCREENSHOT_FILE". (Read-only)}

\item{firmwareVersion}{The Chrome device's firmware version.}

\item{systemRamFreeReports}{Reports of amounts of available RAM memory (Read-only)}

\item{orderNumber}{The device's order number. Only devices directly purchased from Google have an order number.}

\item{ethernetMacAddress}{The device's MAC address on the ethernet network interface.}

\item{platformVersion}{The Chrome device's platform version.}

\item{deviceFiles}{List of device files to download (Read-only)}

\item{macAddress}{The device's wireless MAC address. If the device does not have this information, it is not included in the response.}

\item{willAutoRenew}{Determines if the device will auto renew its support after the support end date. This is a read-only property.}

\item{recentUsers}{List of recent device users, in descending order, by last login time.}

\item{annotatedLocation}{The address or location of the device as noted by the administrator. Maximum length is \code{200} characters. Empty values are allowed.}

\item{systemRamTotal}{Total RAM on the device \link{in bytes} (Read-only)}

\item{status}{The status of the device.}

\item{annotatedUser}{The user of the device as noted by the administrator. Maximum length is 100 characters. Empty values are allowed.}

\item{model}{The device's model information. If the device does not have this information, this property is not included in the response.}

\item{supportEndDate}{Final date the device will be supported (Read-only)}

\item{ethernetMacAddress0}{(Read-only) MAC address used by the Chromebook’s internal ethernet port, and for onboard network (ethernet) interface. The format is twelve (12) hexadecimal digits without any delimiter (uppercase letters). This is only relevant for some devices.}

\item{notes}{Notes about this device added by the administrator. This property can be \href{https://support.google.com/chrome/a/answer/1698333}{searched} with the \href{/admin-sdk/directory/v1/reference/chromeosdevices/list}{list} method's \code{query} parameter. Maximum length is 500 characters. Empty values are allowed.}

\item{serialNumber}{The Chrome device serial number entered when the device was enabled. This value is the same as the Admin console's \emph{Serial Number} in the \emph{Chrome OS Devices} tab.}

\item{diskVolumeReports}{Reports of disk space and other info about mounted/connected volumes.}

\item{cpuStatusReports}{Reports of CPU utilization and temperature (Read-only)}

\item{bootMode}{The boot mode for the device. The possible values are: * \code{Verified}: The device is running a valid version of the Chrome OS. * \code{Dev}: The devices's developer hardware switch is enabled. When booted, the device has a command line shell. For an example of a developer switch, see the \href{https://www.chromium.org/chromium-os/developer-information-for-chrome-os-devices/samsung-series-5-chromebook#TOC-Developer-switch}{Chromebook developer information}.}

\item{orgUnitPath}{The full parent path with the organizational unit's name associated with the device. Path names are case insensitive. If the parent organizational unit is the top-level organization, it is represented as a forward slash, \code{/}. This property can be \href{/admin-sdk/directory/v1/guides/manage-chrome-devices#update_chrome_device}{updated} using the API. For more information about how to create an organizational structure for your device, see the \href{https://support.google.com/a/answer/182433}{administration help center}.}

\item{lastSync}{Date and time the device was last synchronized with the policy settings in the G Suite administrator control panel (Read-only)}

\item{annotatedAssetId}{The asset identifier as noted by an administrator or specified during enrollment.}

\item{osVersion}{The Chrome device's operating system version.}

\item{lastEnrollmentTime}{Date and time the device was last enrolled (Read-only)}

\item{kind}{The type of resource. For the Chromeosdevices resource, the value is \code{admin#directory#chromeosdevice}.}

\item{lastKnownNetwork}{Contains last known network (Read-only)}

\item{activeTimeRanges}{List of active time ranges (Read-only).}

\item{manufactureDate}{(Read-only) The date the device was manufactured in yyyy-mm-dd format.}

\item{tpmVersionInfo}{Trusted Platform Module (TPM) (Read-only)}

\item{meid}{The Mobile Equipment Identifier (MEID) or the International Mobile Equipment Identity (IMEI) for the 3G mobile card in a mobile device. A MEID/IMEI is typically used when adding a device to a wireless carrier's post-pay service plan. If the device does not have this information, this property is not included in the response. For more information on how to export a MEID/IMEI list, see the \href{/admin-sdk/directory/v1/guides/manage-chrome-devices.html#export_meid}{Developer's Guide}.}

\item{dockMacAddress}{(Read-only) Built-in MAC address for the docking station that the device connected to. Factory sets Media access control address (MAC address) assigned for use by a dock. It is reserved specifically for MAC pass through device policy. The format is twelve (12) hexadecimal digits without any delimiter (uppercase letters). This is only relevant for some devices.}

\item{etag}{ETag of the resource.}

\item{autoUpdateExpiration}{(Read-only) The timestamp after which the device will stop receiving Chrome updates or support}

\item{callback}{JSONP}

\item{prettyPrint}{Returns response with indentations and line breaks.}

\item{alt}{Data format for response.}

\item{quotaUser}{Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.}

\item{fields}{Selector specifying which fields to include in a partial response.}

\item{uploadType}{Legacy upload protocol for media (e.g. "media", "multipart").}

\item{upload_protocol}{Upload protocol for media (e.g. "raw", "multipart").}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{package_make}}.
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.device.chromeos
}
}
\seealso{
\href{https://developers.google.com/admin-sdk/}{Google Documentation}
}
