% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customers.R
\name{customers.update}
\alias{customers.update}
\title{Updates a customer.}
\usage{
customers.update(
  customerKey,
  phoneNumber = NULL,
  customerCreationTime = NULL,
  id = NULL,
  alternateEmail = NULL,
  language = NULL,
  etag = NULL,
  postalAddress = NULL,
  kind = NULL,
  customerDomain = NULL,
  callback = NULL,
  prettyPrint = NULL,
  alt = NULL,
  quotaUser = NULL,
  fields = NULL,
  uploadType = NULL,
  upload_protocol = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.customers.update"]]$scopes[[1]]),
  return_response = F
)
}
\arguments{
\item{customerKey}{Id of the customer to be updated}

\item{phoneNumber}{The customer's contact phone number in \href{https://en.wikipedia.org/wiki/E.164}{E.164} format.}

\item{customerCreationTime}{The customer's creation time (Readonly)}

\item{id}{The unique ID for the customer's Google Workspace account. (Readonly)}

\item{alternateEmail}{The customer's secondary contact email address. This email address cannot be on the same domain as the \code{customerDomain}}

\item{language}{The customer's ISO 639-2 language code. See the \href{/admin-sdk/directory/v1/languages}{Language Codes} page for the list of supported codes. Valid language codes outside the supported set will be accepted by the API but may lead to unexpected behavior. The default value is \code{en}.}

\item{etag}{ETag of the resource.}

\item{postalAddress}{The customer's postal address information.}

\item{kind}{Identifies the resource as a customer. Value: \code{admin#directory#customer}}

\item{customerDomain}{The customer's primary domain name string. Do not include the \code{www} prefix when creating a new customer.}

\item{callback}{JSONP}

\item{prettyPrint}{Returns response with indentations and line breaks.}

\item{alt}{Data format for response.}

\item{quotaUser}{Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.}

\item{fields}{Selector specifying which fields to include in a partial response.}

\item{uploadType}{Legacy upload protocol for media (e.g. "media", "multipart").}

\item{upload_protocol}{Upload protocol for media (e.g. "raw", "multipart").}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{package_make}}.
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.customer
}
}
\seealso{
\href{https://developers.google.com/admin-sdk/}{Google Documentation}
}
