% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domainAliases.R
\name{domainAliases.insert}
\alias{domainAliases.insert}
\title{Inserts a domain alias of the customer.}
\usage{
domainAliases.insert(
  customer,
  kind = NULL,
  verified = NULL,
  creationTime = NULL,
  parentDomainName = NULL,
  domainAliasName = NULL,
  etag = NULL,
  callback = NULL,
  prettyPrint = NULL,
  alt = NULL,
  quotaUser = NULL,
  fields = NULL,
  uploadType = NULL,
  upload_protocol = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.domainAliases.insert"]]$scopes[[1]]),
  return_response = F
)
}
\arguments{
\item{customer}{Immutable ID of the Google Workspace account.}

\item{kind}{Kind of resource this is.}

\item{verified}{Indicates the verification state of a domain alias. (Read-only)}

\item{creationTime}{The creation time of the domain alias. (Read-only).}

\item{parentDomainName}{The parent domain name that the domain alias is associated with. This can either be a primary or secondary domain name within a customer.}

\item{domainAliasName}{The domain alias name.}

\item{etag}{ETag of the resource.}

\item{callback}{JSONP}

\item{prettyPrint}{Returns response with indentations and line breaks.}

\item{alt}{Data format for response.}

\item{quotaUser}{Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.}

\item{fields}{Selector specifying which fields to include in a partial response.}

\item{uploadType}{Legacy upload protocol for media (e.g. "media", "multipart").}

\item{upload_protocol}{Upload protocol for media (e.g. "raw", "multipart").}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{package_make}}.
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.domain
}
}
\seealso{
\href{https://developers.google.com/admin-sdk/}{Google Documentation}
}
