% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domains.R
\name{domains.insert}
\alias{domains.insert}
\title{Inserts a domain of the customer.}
\usage{
domains.insert(
  customer,
  domainAliases = NULL,
  creationTime = NULL,
  verified = NULL,
  domainName = NULL,
  kind = NULL,
  etag = NULL,
  isPrimary = NULL,
  callback = NULL,
  prettyPrint = NULL,
  alt = NULL,
  quotaUser = NULL,
  fields = NULL,
  uploadType = NULL,
  upload_protocol = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.domains.insert"]]$scopes[[1]]),
  return_response = F
)
}
\arguments{
\item{customer}{Immutable ID of the Google Workspace account.}

\item{domainAliases}{List of domain alias objects. (Read-only)}

\item{creationTime}{Creation time of the domain. Expressed in \href{https://en.wikipedia.org/wiki/Epoch_time}{Unix time} format. (Read-only).}

\item{verified}{Indicates the verification state of a domain. (Read-only).}

\item{domainName}{The domain name of the customer.}

\item{kind}{Kind of resource this is.}

\item{etag}{ETag of the resource.}

\item{isPrimary}{Indicates if the domain is a primary domain (Read-only).}

\item{callback}{JSONP}

\item{prettyPrint}{Returns response with indentations and line breaks.}

\item{alt}{Data format for response.}

\item{quotaUser}{Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.}

\item{fields}{Selector specifying which fields to include in a partial response.}

\item{uploadType}{Legacy upload protocol for media (e.g. "media", "multipart").}

\item{upload_protocol}{Upload protocol for media (e.g. "raw", "multipart").}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{package_make}}.
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.domain
}
}
\seealso{
\href{https://developers.google.com/admin-sdk/}{Google Documentation}
}
