% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups.R
\name{groups.update}
\alias{groups.update}
\title{Updates a group's properties.}
\usage{
groups.update(
  groupKey,
  description = NULL,
  nonEditableAliases = NULL,
  directMembersCount = NULL,
  id = NULL,
  name = NULL,
  kind = NULL,
  email = NULL,
  adminCreated = NULL,
  etag = NULL,
  aliases = NULL,
  callback = NULL,
  prettyPrint = NULL,
  alt = NULL,
  quotaUser = NULL,
  fields = NULL,
  uploadType = NULL,
  upload_protocol = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.groups.update"]]$scopes[[1]]),
  return_response = F
)
}
\arguments{
\item{groupKey}{Identifies the group in the API request. The value can be the group's email address, group alias, or the unique group ID.}

\item{description}{An extended description to help users determine the purpose of a group. For example, you can include information about who should join the group, the types of messages to send to the group, links to FAQs about the group, or related groups. Maximum length is \verb{4,096} characters.}

\item{nonEditableAliases}{List of the group's non-editable alias email addresses that are outside of the account's primary domain or subdomains. These are functioning email addresses used by the group. This is a read-only property returned in the API's response for a group. If edited in a group's POST or PUT request, the edit is ignored by the API service.}

\item{directMembersCount}{The number of users that are direct members of the group. If a group is a member (child) of this group (the parent), members of the child group are not counted in the \code{directMembersCount} property of the parent group.}

\item{id}{The unique ID of a group. A group \code{id} can be used as a group request URI's \code{groupKey}.}

\item{name}{The group's display name.}

\item{kind}{The type of the API resource. For Groups resources, the value is \code{admin#directory#group}.}

\item{email}{The group's email address. If your account has multiple domains, select the appropriate domain for the email address. The \code{email} must be unique. This property is required when creating a group. Group email addresses are subject to the same character usage rules as usernames, see the \href{https://support.google.com/a/answer/9193374}{help center} for details.}

\item{adminCreated}{Value is \code{true} if this group was created by an administrator rather than a user.}

\item{etag}{ETag of the resource.}

\item{aliases}{List of a group's alias email addresses.}

\item{callback}{JSONP}

\item{prettyPrint}{Returns response with indentations and line breaks.}

\item{alt}{Data format for response.}

\item{quotaUser}{Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.}

\item{fields}{Selector specifying which fields to include in a partial response.}

\item{uploadType}{Legacy upload protocol for media (e.g. "media", "multipart").}

\item{upload_protocol}{Upload protocol for media (e.g. "raw", "multipart").}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{package_make}}.
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.group
}
}
\seealso{
\href{https://developers.google.com/admin-sdk/}{Google Documentation}
}
