% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/members.R
\name{members.get}
\alias{members.get}
\title{Retrieves a group member's properties.}
\usage{
members.get(
  memberKey,
  groupKey,
  callback = NULL,
  prettyPrint = NULL,
  alt = NULL,
  quotaUser = NULL,
  fields = NULL,
  uploadType = NULL,
  upload_protocol = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.members.get"]]$scopes[[1]]),
  return_response = F
)
}
\arguments{
\item{memberKey}{Identifies the group member in the API request. A group member can be a user or another group. The value can be the member's (group or user) primary email address, alias, or unique ID.}

\item{groupKey}{Identifies the group in the API request. The value can be the group's email address, group alias, or the unique group ID.}

\item{callback}{JSONP}

\item{prettyPrint}{Returns response with indentations and line breaks.}

\item{alt}{Data format for response.}

\item{quotaUser}{Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.}

\item{fields}{Selector specifying which fields to include in a partial response.}

\item{uploadType}{Legacy upload protocol for media (e.g. "media", "multipart").}

\item{upload_protocol}{Upload protocol for media (e.g. "raw", "multipart").}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{package_make}}.
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.group
\item https://www.googleapis.com/auth/admin.directory.group.member
\item https://www.googleapis.com/auth/admin.directory.group.member.readonly
\item https://www.googleapis.com/auth/admin.directory.group.readonly
}
}
\seealso{
\href{https://developers.google.com/admin-sdk/}{Google Documentation}
}
