% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/members.R
\name{members.list}
\alias{members.list}
\title{Retrieves a paginated list of all members in a group.}
\usage{
members.list(
  groupKey,
  includeDerivedMembership = NULL,
  roles = NULL,
  pageToken = NULL,
  maxResults = NULL,
  callback = NULL,
  fields = NULL,
  uploadType = NULL,
  upload_protocol = NULL,
  prettyPrint = NULL,
  quotaUser = NULL,
  alt = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.members.list"]]$scopes[[1]]),
  return_request = F,
  return_response = F
)
}
\arguments{
\item{groupKey}{Identifies the group in the API request. The value can be the group's email address, group alias, or the unique group ID.}

\item{includeDerivedMembership}{Whether to list indirect memberships. Default: false.}

\item{roles}{The \code{roles} query parameter allows you to retrieve group members by role. Allowed values are \code{OWNER}, \code{MANAGER}, and \code{MEMBER}.}

\item{pageToken}{Token to specify next page in the list.}

\item{maxResults}{Maximum number of results to return. Max allowed value is 200.}

\item{callback}{JSONP}

\item{fields}{Selector specifying which fields to include in a partial response.}

\item{uploadType}{Legacy upload protocol for media (e.g. "media", "multipart").}

\item{upload_protocol}{Upload protocol for media (e.g. "raw", "multipart").}

\item{prettyPrint}{Returns response with indentations and line breaks.}

\item{quotaUser}{Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.}

\item{alt}{Data format for response.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_request}{Whether to return the request without making it. Defaults to FALSE. See \code{\link[gargle]{request_build}} for more info on the returned request object.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{package_make}}.
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.group
\item https://www.googleapis.com/auth/admin.directory.group.member
\item https://www.googleapis.com/auth/admin.directory.group.member.readonly
\item https://www.googleapis.com/auth/admin.directory.group.readonly
}
}
\seealso{
\href{https://developers.google.com/admin-sdk/}{Google Documentation}
}
