% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobiledevices.R
\name{mobiledevices.list}
\alias{mobiledevices.list}
\title{Retrieves a paginated list of all mobile devices for an account.}
\usage{
mobiledevices.list(
  customerId,
  query = NULL,
  orderBy = NULL,
  pageToken = NULL,
  projection = NULL,
  sortOrder = NULL,
  maxResults = NULL,
  callback = NULL,
  fields = NULL,
  uploadType = NULL,
  upload_protocol = NULL,
  prettyPrint = NULL,
  quotaUser = NULL,
  alt = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.mobiledevices.list"]]$scopes[[1]]),
  return_request = F,
  return_response = F
)
}
\arguments{
\item{customerId}{The unique ID for the customer's Google Workspace account. As an account administrator, you can also use the \code{my_customer} alias to represent your account's \code{customerId}. The \code{customerId} is also returned as part of the \href{/admin-sdk/directory/v1/reference/users}{Users resource}.}

\item{query}{Search string in the format given at https://developers.google.com/admin-sdk/directory/v1/search-operators}

\item{orderBy}{Device property to use for sorting results.}

\item{pageToken}{Token to specify next page in the list}

\item{projection}{Restrict information returned to a set of selected fields.}

\item{sortOrder}{Whether to return results in ascending or descending order. Must be used with the \code{orderBy} parameter.}

\item{maxResults}{Maximum number of results to return. Max allowed value is 100.}

\item{callback}{JSONP}

\item{fields}{Selector specifying which fields to include in a partial response.}

\item{uploadType}{Legacy upload protocol for media (e.g. "media", "multipart").}

\item{upload_protocol}{Upload protocol for media (e.g. "raw", "multipart").}

\item{prettyPrint}{Returns response with indentations and line breaks.}

\item{quotaUser}{Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.}

\item{alt}{Data format for response.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_request}{Whether to return the request without making it. Defaults to FALSE. See \code{\link[gargle]{request_build}} for more info on the returned request object.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{package_make}}.
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.device.mobile
\item https://www.googleapis.com/auth/admin.directory.device.mobile.action
\item https://www.googleapis.com/auth/admin.directory.device.mobile.readonly
}
}
\seealso{
\href{https://developers.google.com/admin-sdk/}{Google Documentation}
}
