% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orgunits.R
\name{orgunits.patch}
\alias{orgunits.patch}
\title{Updates an organizational unit.}
\usage{
orgunits.patch(
  orgUnitPath,
  customerId,
  name = NULL,
  blockInheritance = NULL,
  parentOrgUnitId = NULL,
  orgUnitId = NULL,
  kind = NULL,
  etag = NULL,
  description = NULL,
  parentOrgUnitPath = NULL,
  callback = NULL,
  prettyPrint = NULL,
  alt = NULL,
  quotaUser = NULL,
  fields = NULL,
  uploadType = NULL,
  upload_protocol = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.orgunits.patch"]]$scopes[[1]]),
  return_response = F
)
}
\arguments{
\item{orgUnitPath}{The full path of the organizational unit or its unique ID.}

\item{customerId}{The unique ID for the customer's Google Workspace account. As an account administrator, you can also use the \code{my_customer} alias to represent your account's \code{customerId}. The \code{customerId} is also returned as part of the \href{/admin-sdk/directory/v1/reference/users}{Users resource}.}

\item{name}{The organizational unit's path name. For example, an organizational unit's name within the /corp/support/sales_support parent path is sales_support. Required.}

\item{blockInheritance}{Determines if a sub-organizational unit can inherit the settings of the parent organization. The default value is \code{false}, meaning a sub-organizational unit inherits the settings of the nearest parent organizational unit. For more information on inheritance and users in an organization structure, see the \href{https://support.google.com/a/answer/4352075}{administration help center}.}

\item{parentOrgUnitId}{The unique ID of the parent organizational unit. Required, unless \code{parentOrgUnitPath} is set.}

\item{orgUnitId}{The unique ID of the organizational unit.}

\item{kind}{The type of the API resource. For Orgunits resources, the value is \code{admin#directory#orgUnit}.}

\item{etag}{ETag of the resource.}

\item{description}{Description of the organizational unit.}

\item{parentOrgUnitPath}{The organizational unit's parent path. For example, /corp/sales is the parent path for /corp/sales/sales_support organizational unit. Required, unless \code{parentOrgUnitId} is set.}

\item{callback}{JSONP}

\item{prettyPrint}{Returns response with indentations and line breaks.}

\item{alt}{Data format for response.}

\item{quotaUser}{Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.}

\item{fields}{Selector specifying which fields to include in a partial response.}

\item{uploadType}{Legacy upload protocol for media (e.g. "media", "multipart").}

\item{upload_protocol}{Upload protocol for media (e.g. "raw", "multipart").}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
This method supports \href{/admin-sdk/directory/v1/guides/performance#patch}{patch semantics} Autogenerated via \code{\link[googlePackageMaker]{package_make}}.
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.orgunit
}
}
\seealso{
\href{https://developers.google.com/admin-sdk/}{Google Documentation}
}
