% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roles.R
\name{roles.update}
\alias{roles.update}
\title{Updates a role.}
\usage{
roles.update(
  customer,
  roleId,
  isSystemRole = NULL,
  kind = NULL,
  etag = NULL,
  roleDescription = NULL,
  isSuperAdminRole = NULL,
  rolePrivileges = NULL,
  roleName = NULL,
  callback = NULL,
  prettyPrint = NULL,
  alt = NULL,
  quotaUser = NULL,
  fields = NULL,
  uploadType = NULL,
  upload_protocol = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.roles.update"]]$scopes[[1]]),
  return_response = F
)
}
\arguments{
\item{customer}{Immutable ID of the Google Workspace account.}

\item{roleId}{Immutable ID of the role.}

\item{isSystemRole}{Returns \code{true} if this is a pre-defined system role.}

\item{kind}{The type of the API resource. This is always \code{admin#directory#role}.}

\item{etag}{ETag of the resource.}

\item{roleDescription}{A short description of the role.}

\item{isSuperAdminRole}{Returns \code{true} if the role is a super admin role.}

\item{rolePrivileges}{The set of privileges that are granted to this role.}

\item{roleName}{Name of the role.}

\item{callback}{JSONP}

\item{prettyPrint}{Returns response with indentations and line breaks.}

\item{alt}{Data format for response.}

\item{quotaUser}{Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.}

\item{fields}{Selector specifying which fields to include in a partial response.}

\item{uploadType}{Legacy upload protocol for media (e.g. "media", "multipart").}

\item{upload_protocol}{Upload protocol for media (e.g. "raw", "multipart").}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{package_make}}.
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.rolemanagement
}
}
\seealso{
\href{https://developers.google.com/admin-sdk/}{Google Documentation}
}
