% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{schemas.insert}
\alias{schemas.insert}
\title{Creates a schema.}
\usage{
schemas.insert(
  customerId,
  kind = NULL,
  fields = NULL,
  displayName = NULL,
  etag = NULL,
  schemaId = NULL,
  schemaName = NULL,
  callback = NULL,
  prettyPrint = NULL,
  alt = NULL,
  quotaUser = NULL,
  uploadType = NULL,
  upload_protocol = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.schemas.insert"]]$scopes[[1]]),
  return_response = F
)
}
\arguments{
\item{customerId}{Immutable ID of the Google Workspace account.}

\item{kind}{Kind of resource this is.}

\item{fields}{A list of fields in the schema.}

\item{displayName}{Display name for the schema.}

\item{etag}{The ETag of the resource.}

\item{schemaId}{The unique identifier of the schema (Read-only)}

\item{schemaName}{The schema's name.}

\item{callback}{JSONP}

\item{prettyPrint}{Returns response with indentations and line breaks.}

\item{alt}{Data format for response.}

\item{quotaUser}{Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.}

\item{uploadType}{Legacy upload protocol for media (e.g. "media", "multipart").}

\item{upload_protocol}{Upload protocol for media (e.g. "raw", "multipart").}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{package_make}}.
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.userschema
}
}
\seealso{
\href{https://developers.google.com/admin-sdk/}{Google Documentation}
}
