% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schemas.R
\name{schemas.list}
\alias{schemas.list}
\title{Retrieves all schemas for a customer.}
\usage{
schemas.list(
  customerId,
  callback = NULL,
  fields = NULL,
  uploadType = NULL,
  upload_protocol = NULL,
  prettyPrint = NULL,
  quotaUser = NULL,
  alt = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["directory.schemas.list"]]$scopes[[1]]),
  return_request = F,
  return_response = F
)
}
\arguments{
\item{customerId}{Immutable ID of the Google Workspace account.}

\item{callback}{JSONP}

\item{fields}{Selector specifying which fields to include in a partial response.}

\item{uploadType}{Legacy upload protocol for media (e.g. "media", "multipart").}

\item{upload_protocol}{Upload protocol for media (e.g. "raw", "multipart").}

\item{prettyPrint}{Returns response with indentations and line breaks.}

\item{quotaUser}{Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.}

\item{alt}{Data format for response.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_request}{Whether to return the request without making it. Defaults to FALSE. See \code{\link[gargle]{request_build}} for more info on the returned request object.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{package_make}}.
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.userschema
\item https://www.googleapis.com/auth/admin.directory.userschema.readonly
}
}
\seealso{
\href{https://developers.google.com/admin-sdk/}{Google Documentation}
}
