% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channels.R
\name{channels.stop}
\alias{channels.stop}
\title{Stops watching resources through this channel.}
\usage{
channels.stop(
  params = NULL,
  expiration = NULL,
  payload = NULL,
  type = NULL,
  kind = NULL,
  address = NULL,
  resourceUri = NULL,
  token = NULL,
  id = NULL,
  resourceId = NULL,
  callback = NULL,
  fields = NULL,
  uploadType = NULL,
  upload_protocol = NULL,
  prettyPrint = NULL,
  quotaUser = NULL,
  alt = NULL,
  gargle_token = gargle::token_fetch(scopes =
    .endpoints[["admin.channels.stop"]]$scopes[[1]]),
  return_request = F,
  return_response = F
)
}
\arguments{
\item{params}{Additional parameters controlling delivery channel behavior. Optional.}

\item{expiration}{Date and time of notification channel expiration, expressed as a Unix timestamp, in milliseconds. Optional.}

\item{payload}{A Boolean value to indicate whether payload is wanted. Optional.}

\item{type}{The type of delivery mechanism used for this channel.}

\item{kind}{Identifies this as a notification channel used to watch for changes to a resource, which is \code{api#channel}.}

\item{address}{The address where notifications are delivered for this channel.}

\item{resourceUri}{A version-specific identifier for the watched resource.}

\item{token}{An arbitrary string delivered to the target address with each notification delivered over this channel. Optional.}

\item{id}{A UUID or similar unique string that identifies this channel.}

\item{resourceId}{An opaque ID that identifies the resource being watched on this channel. Stable across different API versions.}

\item{callback}{JSONP}

\item{fields}{Selector specifying which fields to include in a partial response.}

\item{uploadType}{Legacy upload protocol for media (e.g. "media", "multipart").}

\item{upload_protocol}{Upload protocol for media (e.g. "raw", "multipart").}

\item{prettyPrint}{Returns response with indentations and line breaks.}

\item{quotaUser}{Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters.}

\item{alt}{Data format for response.}

\item{gargle_token}{A token prepared by one of gargle's token generating functions. Defaults to gargle::token_fetch(...) with appropriate scopes. See \code{\link[gargle]{token_fetch}} for more info.}

\item{return_request}{Whether to return the request without making it. Defaults to FALSE. See \code{\link[gargle]{request_build}} for more info on the returned request object.}

\item{return_response}{Whether to return the response or the response content. Defaults to FALSE (return response content).}
}
\description{
Autogenerated via \code{\link[googlePackageMaker]{package_make}}.
}
\details{
Authentication scopes used by this function are:
\itemize{
\item https://www.googleapis.com/auth/admin.directory.user
\item https://www.googleapis.com/auth/admin.directory.user.alias
\item https://www.googleapis.com/auth/admin.directory.user.alias.readonly
\item https://www.googleapis.com/auth/admin.directory.user.readonly
\item https://www.googleapis.com/auth/cloud-platform
}
}
\seealso{
\href{https://developers.google.com/admin-sdk/}{Google Documentation}
}
